// src/app/core/models/user.ts
import { Injectable } from '@angular/core';
export interface User {
    Uid: string;
    Email: string;
    Url?: string;
    Name: string;
    Country: string;
    Sex: string;
    Birth: any;
    Telephone: string;
    AllowFriendSeeLists: any;
    Currency: any;
    Language: any;
    PushNotifications: any;
}